backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.40.64",
    "fastapi>=0.120.4",
    "langchain>=1.0.3",
    "langchain-aws>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-xai>=1.0.0",
    "langgraph>=1.0.2",
    "pypdf>=6.1.3",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "uvicorn>=0.38.0",
    "websockets>=15.0.1",
]

[project.scripts]
main = "my_project.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]
